container_api_raw functions
