+++ clmctfy/clmctfy_container_api.cc
@@ -11,6 +11,7 @@
 #include "clmctfy_macros.h"
 #include "clmctfy_container_struct.h"
 #include "clmctfy_container_api_struct.h"
+#include "clmctfy_container_api_raw.h"
 
 #define STATUS_OK UTIL__ERROR__CODE__OK
 
@@ -22,17 +23,6 @@ using ::util::internal::status_copy;
 using ::util::internal::status_new;
 using ::util::Status;
 using ::util::StatusOr;
-
-int lmctfy_init_machine_raw(const void *spec, const size_t spec_size, struct status *s) {
-  InitSpec init_spec;
-  CHECK_NOTFAIL_OR_RETURN(s);
-  if (spec != NULL && spec_size > 0) {
-    // XXX should we consider this as an error?
-    init_spec.ParseFromArray(spec, spec_size);
-  }
-  Status v = ContainerApi::InitMachine(init_spec);
-  return status_copy(s, v);
-}
 
 int lmctfy_init_machine(const Containers__Lmctfy__InitSpec *spec, struct status *s) {
   uint8_t *buf = NULL;
@@ -93,29 +83,6 @@ int lmctfy_container_api_get_container(const struct container_api *api,
   Container *ctnr = NULL;
 
   StatusOr<Container *> statusor = api->container_api_->Get(container_name);
-  RETURN_IF_ERROR_PTR(s, statusor, &ctnr);
-  COPY_CONTAINER_STRUCTURE(ctnr, c);
-  return STATUS_OK;
-}
-
-int lmctfy_container_api_create_container_raw(struct container_api *api,
-                                       const char *container_name,
-                                       const void *spec,
-                                       const size_t spec_size,
-                                       struct container **c,
-                                       struct status *s) {
-  CHECK_NOTNULL_OR_RETURN(s, api);
-  CHECK_NOTNULL_OR_RETURN(s, api->container_api_);
-  CHECK_NOTNULL_OR_RETURN(s, c);
-  CHECK_NOTNULL_OR_RETURN(s, container_name);
-  CHECK_POSITIVE_OR_RETURN(s, strlen(container_name));
-  ContainerSpec container_spec;
-  Container *ctnr = NULL;
-  if (spec != NULL && spec_size > 0) {
-    container_spec.ParseFromArray(spec, spec_size);
-  }
-
-  StatusOr<Container *> statusor = api->container_api_->Create(container_name, container_spec);
   RETURN_IF_ERROR_PTR(s, statusor, &ctnr);
   COPY_CONTAINER_STRUCTURE(ctnr, c);
   return STATUS_OK;
